Refactoring
Refactoring all my sh so this looks like real code.
