+++ Content.Client/_Stalker_EN/Clothing/Systems/HelmetVisorSystem.cs
@@ -1,0 +1,10 @@
+using Content.Shared._Stalker_EN.Clothing;
+using Content.Shared._Stalker_EN.Clothing.Components;
+using Content.Shared.Clothing;
+
+namespace Content.Client._Stalker_EN.Clothing.Systems;
+
+public sealed class HelmetVisorSystem : SharedHelmetVisorSystem
+{
+
+}
