+++ Content.Server/_Stalker_EN/Clothing/HelmetVisorSystem.cs
@@ -1,0 +1,10 @@
+using Content.Shared._Stalker_EN.Clothing;
+using Content.Shared._Stalker_EN.Clothing.Components;
+using Content.Shared.Clothing;
+
+namespace Content.Server._Stalker_EN.Clothing;
+
+public sealed class HelmetVisorSystem : SharedHelmetVisorSystem
+{
+
+}
