+++ Content.Shared/_Stalker_EN/Clothing/Components/SlotBlockOverrideComponent.cs
